Make PPFX affect render texture only so it can be used in RawImage
In the Renderer2D, the render texture of the camera is cleared, and the additional render targets are cleared too. The render target to use in the custom PPFXs depends on the previous PP step.
Custom volume components now have default values that make them not active, so they are not added by default to the VolumeState of the VolumeManager created for the camera.
In the custom PPFX render passes they now decide which render target to use, depending on whether the camera uses a render texture or not.
Now the screen border PPFX shader is transparent and the central square is not discarded but written with 0 alpha, so it can be merged later atop other images.
The uber post have been modified so it writes the actual resulting alpha channel, instead of opaque. The bloom FX is now transparent so the color it irradiates can be blended with other images drawn before to the backbuffer.