Optimization: Lights texture caching implemented
Now when static lights are calculated and drawn on the temporary render texture that is used as input when drawing the sprites, the resultant pixels can be copied to an external render texture to be stored.
Once we have some cached textures, it's possible to provide such textures to the renderer so, instead of calculating the static lights and shadows again, the cached result stored in the external textures is copied to the temporary render texture, saving a potentially huge amount of GPU usage.
RendererData asset is now accessible through the ScriptableRenderer.
Added 2 new shaders.
Added a RenderTexture asset.