Fixed: Static lights are not smoothed
When the dynamic lights were going to be rendered, the "has rendered shadows" flag was reset to false, so if there was not any dynamic light for a sorting layer, the renderer was thinking that shadow smoothing was not necessary. The flag coming from rendering static lights is now taking into account.
Now it is possible to specify whether a cached light texture contains shadows or not, so the renderer knows if shadow smoothing is needed.